Revert " ♻️ Refactor: unify common components Part 1 #1029"
This reverts commit 4d045e5f531be824062c665bf33bcbb3c5b23ea6, reversing
changes made to 89fd7bea080e66ecc9afd6dfe0c913793ad25faf.
