+++ frontend/app/[locale]/chat/internal/chatInterface.tsx
@@ -2,33 +2,33 @@
 
 import type React from "react";
 
-import { useState, useEffect } from "react";
+import { useState, useRef, useEffect } from "react";
 import { useRouter } from "next/navigation";
 import { v4 as uuidv4 } from "uuid";
 import { useTranslation } from "react-i18next";
 
 import { ROLE_ASSISTANT } from "@/const/agentConfig";
+import { chatConfig } from "@/const/chatConfig";
 import { USER_ROLES } from "@/const/modelConfig";
 import { useConfig } from "@/hooks/useConfig";
 import { useAuth } from "@/hooks/useAuth";
 import { conversationService } from "@/services/conversationService";
-import { useConversationManagement } from "@/hooks/chat/useConversationManagement";
-import { useMessageManagement } from "@/hooks/chat/useMessageManagement";
-import { useAttachmentHandlers } from "@/hooks/chat/useAttachmentHandlers";
-import { useChatUIState } from "@/hooks/chat/useChatUIState";
+import { storageService } from "@/services/storageService";
 
 import { ChatSidebar } from "../components/chatLeftSidebar";
+import { FilePreview } from "@/types/chat";
 import { ChatHeader } from "../components/chatHeader";
 import { ChatRightPanel } from "../components/chatRightPanel";
 import { ChatStreamMain } from "../streaming/chatStreamMain";
 
 import {
   preprocessAttachments,
-  createPreprocessingMessage,
-  updatePreprocessingStep,
-  handlePreprocessingError,
+  handleFileUpload as preProcessHandleFileUpload,
+  handleImageUpload as preProcessHandleImageUpload,
+  uploadAttachments,
+  createMessageAttachments,
+  cleanupAttachmentUrls,
 } from "@/app/chat/internal/chatPreprocess";
-import { loadAttachmentUrls } from "@/app/chat/internal/chatAttachment";
 import {
   Tooltip,
   TooltipContent,
@@ -49,46 +49,161 @@ import log from "@/lib/logger";
 
 const stepIdCounter = { current: 0 };
 
+// Get internationalization key based on message type
+const getI18nKeyByType = (type: string): string => {
+  const typeToKeyMap: Record<string, string> = {
+    "progress": "chatInterface.parsingFileWithProgress",
+    "truncation": "chatInterface.fileTruncated",
+  };
+  return typeToKeyMap[type] || "";
+};
+
 export function ChatInterface() {
   const router = useRouter();
   const { user } = useAuth(); // Get user information
   const [input, setInput] = useState("");
-  
-  // Use message management hook
-  const messageManagement = useMessageManagement();
-  
-  // Use conversation management hook
-  const conversationManagement = useConversationManagement();
-  
-  // Use UI state management hook
-  const uiState = useChatUIState();
-  
+  // Replace the original messages state
+  const [sessionMessages, setSessionMessages] = useState<{
+    [conversationId: number]: ChatMessageType[];
+  }>({});
+  const [isSwitchedConversation, setIsSwitchedConversation] = useState(false); // Add conversation switching tracking state
+  const [isLoading, setIsLoading] = useState(false);
+  const initialized = useRef(false);
   const { t } = useTranslation("common");
+  const [conversationTitle, setConversationTitle] = useState(
+    t("chatInterface.newConversation")
+  );
+  const [conversationId, setConversationId] = useState<number>(0);
+  const [conversationList, setConversationList] = useState<
+    ConversationListItem[]
+  >([]);
+  const [selectedConversationId, setSelectedConversationId] = useState<
+    number | null
+  >(null);
+  const [openDropdownId, setOpenDropdownId] = useState<string | null>(null);
   const { appConfig } = useConfig();
 
-  // Use attachment handlers hook
-  const attachmentHandlers = useAttachmentHandlers();
+  // For each conversation, maintain independent SSE connections and states
+  const [streamingConversations, setStreamingConversations] = useState<
+    Set<number>
+  >(new Set());
+  const conversationControllersRef = useRef<Map<number, AbortController>>(
+    new Map()
+  );
+  const conversationTimeoutsRef = useRef<Map<number, NodeJS.Timeout>>(
+    new Map()
+  );
 
   // Place the declaration of currentMessages after the definition of selectedConversationId
   // If a historical conversation is being loaded and there are no cached messages, return an empty array to avoid displaying error content
-  const currentMessages = conversationManagement.selectedConversationId
-    ? messageManagement.getMessages(conversationManagement.selectedConversationId)
+  const currentMessages = selectedConversationId
+    ? sessionMessages[selectedConversationId] || []
     : [];
 
   // Monitor changes in currentMessages
   // Calculate if the current conversation is streaming
   const isCurrentConversationStreaming =
-    conversationManagement.conversationId && conversationManagement.conversationId !== -1
-      ? uiState.streamingConversations.has(conversationManagement.conversationId)
+    conversationId && conversationId !== -1
+      ? streamingConversations.has(conversationId)
       : false;
 
+  const [viewingImage, setViewingImage] = useState<string | null>(null);
+
+  // Add attachment state management
+  const [attachments, setAttachments] = useState<FilePreview[]>([]);
+  const [fileUrls, setFileUrls] = useState<{ [id: string]: string }>({});
+
+  const [isStreaming, setIsStreaming] = useState(false); // Add streaming state
+  const abortControllerRef = useRef<AbortController | null>(null); // Add AbortController reference
+  const timeoutRef = useRef<NodeJS.Timeout | null>(null); // Add timeout reference
+
+  // Add sidebar state control
+  const [sidebarOpen, setSidebarOpen] = useState(true);
+
+  // Add a new state for new conversation status
+  const [isNewConversation, setIsNewConversation] = useState(true);
+
+  // Add a state to track if we're loading a historical conversation
+  const [isLoadingHistoricalConversation, setIsLoadingHistoricalConversation] =
+    useState(false);
+
+  // Add a state to track conversation loading errors
+  const [conversationLoadError, setConversationLoadError] = useState<{
+    [conversationId: number]: string;
+  }>({});
+
+  // Add a state to track completed conversations that haven't been viewed yet
+  const [completedConversations, setCompletedConversations] = useState<
+    Set<number>
+  >(new Set());
+
+  // Add a ref to track the currently selected conversation ID for real-time access
+  const currentSelectedConversationRef = useRef<number | null>(null);
+
+  // Ensure right sidebar is closed by default
+  const [showRightPanel, setShowRightPanel] = useState(false);
+
+  const [selectedMessageId, setSelectedMessageId] = useState<
+    string | undefined
+  >();
+
+  // Add force scroll to bottom state control
+  const [shouldScrollToBottom, setShouldScrollToBottom] = useState(false);
+
+  // Add agent selection state
+  const [selectedAgentId, setSelectedAgentId] = useState<number | null>(null);
+
+  // Reset scroll to bottom state
+  useEffect(() => {
+    if (shouldScrollToBottom) {
+      // Give enough time for scrolling to complete, then reset state
+      const timer = setTimeout(() => {
+        setShouldScrollToBottom(false);
+      }, 1200); // Slightly longer than the last scroll delay in ChatStreamMain
+
+      return () => clearTimeout(timer);
+    }
+  }, [shouldScrollToBottom]);
+
+  // Add attachment cleanup function - cleanup URLs when component unmounts
+  useEffect(() => {
+    return () => {
+      // Use preprocessing function to cleanup URLs
+      cleanupAttachmentUrls(attachments, fileUrls);
+    };
+  }, [attachments, fileUrls]);
+
+  // Handle file upload
+  const handleFileUpload = (file: File) => {
+    return preProcessHandleFileUpload(file, setFileUrls, t);
+  };
+
+  // Handle image upload
+  const handleImageUpload = (file: File) => {
+    preProcessHandleImageUpload(file, t);
+  };
+
+  // Add attachment management function
+  const handleAttachmentsChange = (newAttachments: FilePreview[]) => {
+    setAttachments(newAttachments);
+  };
+
+  // Define sidebar toggle function
+  const toggleSidebar = () => {
+    setSidebarOpen(!sidebarOpen);
+  };
+
+  // Handle right panel toggle - keep it simple and clear
+  const toggleRightPanel = () => {
+    setShowRightPanel(!showRightPanel);
+  };
 
   useEffect(() => {
-    if (!conversationManagement.initialized.current) {
-      conversationManagement.initialized.current = true;
+    if (!initialized.current) {
+      initialized.current = true;
 
       // Get conversation history list, but don't auto-select the latest conversation
-      conversationManagement.fetchConversationList()
+      fetchConversationList()
         .then((dialogData) => {
           // Create new conversation by default regardless of history
           handleNewConversation();
@@ -101,32 +216,62 @@ export function ChatInterface() {
     }
   }, [appConfig]); // Add appConfig as dependency
 
+  // Add useEffect to listen for conversationId changes, ensure right sidebar is always closed when conversation switches
+  useEffect(() => {
+    // Ensure right sidebar is reset to closed state whenever conversation ID changes
+    setSelectedMessageId(undefined);
+    setShowRightPanel(false);
+  }, [conversationId]);
+
+  // Ensure currentSelectedConversationRef is synchronized with selectedConversationId
+  useEffect(() => {
+    currentSelectedConversationRef.current = selectedConversationId;
+  }, [selectedConversationId]);
+
+  // Clear all timers and requests when component unmounts
+  useEffect(() => {
+    return () => {
+      if (abortControllerRef.current) {
+        try {
+          abortControllerRef.current.abort(t("chatInterface.componentUnmount"));
+        } catch (error) {
+          log.error(t("chatInterface.errorCancelingRequest"), error);
+        }
+        abortControllerRef.current = null;
+      }
+
+      if (timeoutRef.current) {
+        clearTimeout(timeoutRef.current);
+        timeoutRef.current = null;
+      }
+    };
+  }, []);
 
   const handleSend = async () => {
-    if (!input.trim() && attachmentHandlers.attachments.length === 0) return; // Allow sending attachments only, without text content
+    if (!input.trim() && attachments.length === 0) return; // Allow sending attachments only, without text content
 
     // Flag to track if we should reset button states in finally block
     let shouldResetButtonStates = true;
 
     // If in new conversation state, switch to conversation state after sending message
-    if (conversationManagement.isNewConversation) {
-      conversationManagement.setIsNewConversation(false);
+    if (isNewConversation) {
+      setIsNewConversation(false);
     }
 
     // Ensure right sidebar doesn't auto-expand when sending new message
-    uiState.setSelectedMessageId(undefined);
-    uiState.setShowRightPanel(false);
+    setSelectedMessageId(undefined);
+    setShowRightPanel(false);
 
     // Handle user message content
     const userMessageId = uuidv4();
     const userMessageContent = input.trim();
 
     // Get current conversation ID
-    let currentConversationId = conversationManagement.conversationId;
+    let currentConversationId = conversationId;
 
     // Ensure ref reflects the current conversation state
     if (currentConversationId && currentConversationId !== -1) {
-      conversationManagement.currentSelectedConversationRef.current = currentConversationId;
+      currentSelectedConversationRef.current = currentConversationId;
     }
 
     // Prepare attachment information
@@ -134,19 +279,21 @@ export function ChatInterface() {
     let uploadedFileUrls: Record<string, string> = {};
     let objectNames: Record<string, string> = {}; // Add object name mapping
 
-    if (attachmentHandlers.attachments.length > 0) {
+    if (attachments.length > 0) {
       // Show loading state
-      uiState.setIsLoading(true);
+      setIsLoading(true);
 
-      // Use attachment handlers to upload attachments
-      const uploadResult = await attachmentHandlers.uploadAttachmentsToStorage();
+      // Use preprocessing function to upload attachments
+      const uploadResult = await uploadAttachments(attachments, t);
       uploadedFileUrls = uploadResult.uploadedFileUrls;
       objectNames = uploadResult.objectNames; // Get object name mapping
     }
 
-    // Use attachment handlers to create message attachments
-    const messageAttachments = attachmentHandlers.createMessageAttachmentsFromCurrent(
-      uploadedFileUrls
+    // Use preprocessing function to create message attachments
+    const messageAttachments = createMessageAttachments(
+      attachments,
+      uploadedFileUrls,
+      fileUrls
     );
 
     // Create user message object
@@ -161,7 +308,7 @@ export function ChatInterface() {
 
     // Clear input box and attachments
     setInput("");
-    attachmentHandlers.clearAttachments();
+    setAttachments([]);
 
     // Create initial AI reply message
     const assistantMessageId = uuidv4();
@@ -175,10 +322,10 @@ export function ChatInterface() {
     };
 
     // Send message and scroll to bottom
-    uiState.setShouldScrollToBottom(true);
+    setShouldScrollToBottom(true);
 
-    uiState.setIsLoading(true);
-    uiState.setIsStreaming(true); // Set streaming state to true
+    setIsLoading(true);
+    setIsStreaming(true); // Set streaming state to true
 
     // Create independent AbortController for current conversation
     const currentController = new AbortController();
@@ -194,16 +341,16 @@ export function ChatInterface() {
           currentConversationId = createData.conversation_id;
 
           // Update current session state
-          conversationManagement.setConversationId(currentConversationId);
-          conversationManagement.setSelectedConversationId(currentConversationId);
+          setConversationId(currentConversationId);
+          setSelectedConversationId(currentConversationId);
           // Update ref to track current selected conversation
-          conversationManagement.currentSelectedConversationRef.current = currentConversationId;
-          conversationManagement.setConversationTitle(
+          currentSelectedConversationRef.current = currentConversationId;
+          setConversationTitle(
             createData.conversation_title || t("chatInterface.newConversation")
           );
 
           // After creating new conversation, add it to streaming list
-          uiState.setStreamingConversations((prev) => {
+          setStreamingConversations((prev) => {
             const newSet = new Set(prev).add(currentConversationId);
 
             return newSet;
@@ -211,12 +358,12 @@ export function ChatInterface() {
 
           // Refresh conversation list
           try {
-            const dialogList = await conversationManagement.fetchConversationList();
+            const dialogList = await fetchConversationList();
             const newDialog = dialogList.find(
               (dialog) => dialog.conversation_id === currentConversationId
             );
             if (newDialog) {
-              conversationManagement.setSelectedConversationId(currentConversationId);
+              setSelectedConversationId(currentConversationId);
             }
           } catch (error) {
             log.error(
@@ -230,19 +377,19 @@ export function ChatInterface() {
             error
           );
           // Reset button states when conversation creation fails
-          uiState.setIsLoading(false);
-          uiState.setIsStreaming(false);
+          setIsLoading(false);
+          setIsStreaming(false);
           return;
         }
       }
 
       // Ensure valid conversation ID before registering controller and streaming state
       if (currentConversationId && currentConversationId !== -1) {
-        uiState.conversationControllersRef.current.set(
+        conversationControllersRef.current.set(
           currentConversationId,
           currentController
         );
-        uiState.setStreamingConversations((prev) => {
+        setStreamingConversations((prev) => {
           const newSet = new Set(prev);
           newSet.add(currentConversationId);
           return newSet;
@@ -251,36 +398,80 @@ export function ChatInterface() {
 
       // Now add messages after conversation is created/confirmed
       // 1. When sending user message, complete ChatMessageType fields
-      messageManagement.addMessage(currentConversationId, {
-        ...userMessage,
-        id: userMessage.id || uuidv4(),
-        timestamp: userMessage.timestamp || new Date(),
-        isComplete: userMessage.isComplete ?? true,
-        steps: userMessage.steps || [],
-        attachments: userMessage.attachments || [],
-        images: userMessage.images || [],
-      });
+      setSessionMessages((prev) => ({
+        ...prev,
+        [currentConversationId]: [
+          ...(prev[currentConversationId] || []),
+          {
+            ...userMessage,
+            id: userMessage.id || uuidv4(),
+            timestamp: userMessage.timestamp || new Date(),
+            isComplete: userMessage.isComplete ?? true,
+            steps: userMessage.steps || [],
+            attachments: userMessage.attachments || [],
+            images: userMessage.images || [],
+          },
+        ],
+      }));
 
       // 2. When adding AI reply message, complete ChatMessageType fields
-      messageManagement.addMessage(currentConversationId, {
-        ...initialAssistantMessage,
-        id: initialAssistantMessage.id || uuidv4(),
-        timestamp: initialAssistantMessage.timestamp || new Date(),
-        isComplete: initialAssistantMessage.isComplete ?? false,
-        steps: initialAssistantMessage.steps || [],
-        attachments: initialAssistantMessage.attachments || [],
-        images: initialAssistantMessage.images || [],
-      });
+      setSessionMessages((prev) => ({
+        ...prev,
+        [currentConversationId]: [
+          ...(prev[currentConversationId] || []),
+          {
+            ...initialAssistantMessage,
+            id: initialAssistantMessage.id || uuidv4(),
+            timestamp: initialAssistantMessage.timestamp || new Date(),
+            isComplete: initialAssistantMessage.isComplete ?? false,
+            steps: initialAssistantMessage.steps || [],
+            attachments: initialAssistantMessage.attachments || [],
+            images: initialAssistantMessage.images || [],
+          },
+        ],
+      }));
 
       // If there are attachment files, preprocess first
       let finalQuery = userMessage.content;
       // Declare a variable to save file description information
       let fileDescriptionsMap: Record<string, string> = {};
 
-      if (attachmentHandlers.attachments.length > 0) {
+      if (attachments.length > 0) {
         // Attachment preprocessing step, as independent step in assistant steps
-        const preprocessingMessage = createPreprocessingMessage(t);
-        messageManagement.addMessage(currentConversationId, preprocessingMessage);
+        setSessionMessages((prev) => ({
+          ...prev,
+          [currentConversationId]: [
+            ...(prev[currentConversationId] || []),
+            {
+              id: uuidv4(),
+              role: ROLE_ASSISTANT,
+              content: "",
+              timestamp: new Date(),
+              isComplete: false,
+              steps: [
+                {
+                  id: `preprocess-${Date.now()}`,
+                  title: t("chatInterface.filePreprocessing"),
+                  content: "",
+                  expanded: true,
+                  metrics: "",
+                  thinking: { content: "", expanded: false },
+                  code: { content: "", expanded: false },
+                  output: { content: "", expanded: false },
+                  contents: [
+                    {
+                      id: `preprocess-content-${Date.now()}`,
+                      type: chatConfig.contentTypes.PREPROCESS,
+                      content: t("chatInterface.parsingFile"),
+                      expanded: false,
+                      timestamp: Date.now(),
+                    },
+                  ],
+                },
+              ],
+            },
+          ],
+        }));
 
         // Buffer for truncation messages with deduplication
         const truncationBuffer: any[] = [];
@@ -289,17 +480,119 @@ export function ChatInterface() {
         // Use extracted preprocessing function to process attachments
         const result = await preprocessAttachments(
           userMessage.content,
-          attachmentHandlers.attachments,
+          attachments,
           currentController.signal,
           (jsonData) => {
-            uiState.updateLastMessageInConversation(currentConversationId, (lastMsg) => {
-              return updatePreprocessingStep(
-                lastMsg,
-                jsonData,
-                t,
-                truncationBuffer,
-                processedTruncationIds
-              );
+            setSessionMessages((prev) => {
+              const newMessages = { ...prev };
+              const lastMsg =
+                newMessages[currentConversationId]?.[
+                  newMessages[currentConversationId].length - 1
+                ];
+              if (lastMsg && lastMsg.role === ROLE_ASSISTANT) {
+                if (!lastMsg.steps) lastMsg.steps = [];
+                // Find the latest preprocessing step
+                let step = lastMsg.steps.find(
+                  (s) => s.title === t("chatInterface.filePreprocessing")
+                );
+                if (!step) {
+                  step = {
+                    id: `preprocess-${Date.now()}`,
+                    title: t("chatInterface.filePreprocessing"),
+                    content: "",
+                    expanded: true,
+                    metrics: "",
+                    thinking: { content: "", expanded: false },
+                    code: { content: "", expanded: false },
+                    output: { content: "", expanded: false },
+                    contents: [
+                      {
+                        id: `preprocess-content-${Date.now()}`,
+                        type: chatConfig.contentTypes.PREPROCESS,
+                        content: t("chatInterface.parsingFile"),
+                        expanded: false,
+                        timestamp: Date.now(),
+                      },
+                    ],
+                  };
+                  lastMsg.steps.push(step);
+                }
+
+                // Handle truncation messages - buffer them instead of updating immediately
+                if (jsonData.type === "truncation") {
+                  // Create a unique ID for this truncation message to avoid duplicates
+                  const truncationId = `${jsonData.filename || "unknown"}_${
+                    jsonData.message || ""
+                  }`;
+
+                  // Only add if not already processed
+                  if (!processedTruncationIds.has(truncationId)) {
+                    truncationBuffer.push(jsonData);
+                    processedTruncationIds.add(truncationId);
+                  }
+                  return newMessages; // Don't update stepContent for truncation
+                }
+
+                let stepContent = "";
+                switch (jsonData.type) {
+                  case "progress":
+                    if (jsonData.message_data) {
+                      const i18nKey = getI18nKeyByType(jsonData.type);
+                      stepContent = String(
+                        t(i18nKey, jsonData.message_data.params)
+                      );
+                    } else {
+                      stepContent = jsonData.message || "";
+                    }
+                    break;
+                  case "error":
+                    stepContent = t("chatInterface.parseFileFailed", {
+                      filename: jsonData.filename,
+                      message: jsonData.message,
+                    });
+                    break;
+                  case "file_processed":
+                    stepContent = t("chatInterface.fileParsed", {
+                      filename: jsonData.filename,
+                    });
+                    break;
+                  case "complete":
+                    // When complete, process all buffered truncation messages
+                    if (truncationBuffer.length > 0) {
+                      // Process truncation messages using internationalization
+                      const truncationInfo = truncationBuffer
+                        .map((truncation) => {
+                          if (truncation.message_data) {
+                            const i18nKey = getI18nKeyByType(truncation.type);
+                            return String(
+                              t(i18nKey, truncation.message_data.params)
+                            );
+                          } else {
+                            return truncation.message;
+                          }
+                        })
+                        .join(String(t("chatInterface.truncationSeparator")));
+
+                      stepContent = t(
+                        "chatInterface.fileParsingCompleteWithTruncation",
+                        {
+                          truncationInfo: truncationInfo,
+                        }
+                      );
+                    } else {
+                      stepContent = t("chatInterface.fileParsingComplete");
+                    }
+                    break;
+                  default:
+                    stepContent = jsonData.message || "";
+                }
+                // Only update the first content, don't add new ones
+                if (step && step.contents && step.contents.length > 0) {
+                  step.contents[0].content = stepContent;
+                  step.contents[0].timestamp = Date.now();
+                }
+              }
+              return newMessages;
             });
           },
           t,
@@ -309,20 +602,44 @@ export function ChatInterface() {
         // Handle preprocessing result
         if (!result.success) {
           // Reset button states immediately when preprocessing fails
-          uiState.setIsLoading(false);
-          uiState.setIsStreaming(false);
+          setIsLoading(false);
+          setIsStreaming(false);
             
           // Remove from streaming conversations (both new and existing conversations)
           if (currentConversationId) {
-            uiState.setStreamingConversations((prev) => {
+            setStreamingConversations((prev) => {
               const newSet = new Set(prev);
               newSet.delete(currentConversationId);
               return newSet;
             });
           }
           
-          uiState.updateLastMessageInConversation(currentConversationId, (lastMsg) => {
-            return handlePreprocessingError(lastMsg, result.error || "", t);
+          setSessionMessages((prev) => {
+            const newMessages = { ...prev };
+            const lastMsg =
+              newMessages[currentConversationId]?.[
+                newMessages[currentConversationId].length - 1
+              ];
+            
+            if (lastMsg && lastMsg.role === ROLE_ASSISTANT) {
+              // Handle error codes with internationalization
+              let errorMessage;
+              if (result.error === 'REQUEST_ENTITY_TOO_LARGE') {
+                errorMessage = t("chatInterface.fileSizeExceeded");
+              } else if (result.error === 'FILE_PARSING_FAILED') {
+                errorMessage = t("chatInterface.fileParsingFailed");
+              } else {
+                // For any other error, show a generic message without exposing technical details
+                errorMessage = t("chatInterface.fileProcessingFailed", {
+                  error: "Unknown error"
+                });
+              }
+              
+              lastMsg.content = errorMessage;
+              lastMsg.isComplete = true;
+            }
+            
+            return newMessages;
           });
           shouldResetButtonStates = false; // Don't reset again in finally block
           return;
@@ -336,7 +653,7 @@ export function ChatInterface() {
       const runAgentParams: any = {
         query: finalQuery, // Use preprocessed query or original query
         conversation_id: currentConversationId,
-        is_set: uiState.isSwitchedConversation || currentMessages.length <= 1,
+        is_set: isSwitchedConversation || currentMessages.length <= 1,
         history: currentMessages
           .filter((msg) => msg.id !== userMessage.id)
           .map((msg) => ({
@@ -348,7 +665,7 @@ export function ChatInterface() {
           })),
         minio_files:
           messageAttachments.length > 0
-            ? messageAttachments.map((attachment: any) => {
+            ? messageAttachments.map((attachment) => {
                 // Get file description
                 let description = "";
                 if (attachment.name in fileDescriptionsMap) {
@@ -368,8 +685,8 @@ export function ChatInterface() {
       };
 
       // Only add agent_id if it's not null
-      if (uiState.selectedAgentId !== null) {
-        runAgentParams.agent_id = uiState.selectedAgentId;
+      if (selectedAgentId !== null) {
+        runAgentParams.agent_id = selectedAgentId;
       }
 
       const reader = await conversationService.runAgent(
@@ -386,42 +703,52 @@ export function ChatInterface() {
           targetConversationId: number
         ): React.Dispatch<React.SetStateAction<ChatMessageType[]>> =>
         (valueOrUpdater) => {
-          const currentMessages = messageManagement.getMessages(targetConversationId);
-          let nextArr: ChatMessageType[];
-          if (typeof valueOrUpdater === "function") {
-            nextArr = (
-              valueOrUpdater as (prev: ChatMessageType[]) => ChatMessageType[]
-            )(currentMessages);
-          } else {
-            nextArr = valueOrUpdater;
-          }
-          // Update messages using message management
-          messageManagement.setMessages(targetConversationId, nextArr);
+          setSessionMessages((prev) => {
+            const prevArr = prev[targetConversationId] || [];
+            let nextArr: ChatMessageType[];
+            if (typeof valueOrUpdater === "function") {
+              nextArr = (
+                valueOrUpdater as (prev: ChatMessageType[]) => ChatMessageType[]
+              )(prevArr);
+            } else {
+              nextArr = valueOrUpdater;
+            }
+            // Ensure new reference
+            return {
+              ...prev,
+              [targetConversationId]: [...nextArr],
+            };
+          });
         };
 
       // Create resetTimeout function for current conversation
       const resetTimeout = () => {
-        const timeout = uiState.conversationTimeoutsRef.current.get(
+        const timeout = conversationTimeoutsRef.current.get(
           currentConversationId
         );
         if (timeout) {
           clearTimeout(timeout);
         }
         const newTimeout = setTimeout(async () => {
-          const controller = uiState.conversationControllersRef.current.get(
+          const controller = conversationControllersRef.current.get(
             currentConversationId
           );
           if (controller && !controller.signal.aborted) {
             try {
               controller.abort(t("chatInterface.requestTimeout"));
 
-              uiState.updateLastMessageInConversation(currentConversationId, (lastMsg) => {
-                if (lastMsg.role === ROLE_ASSISTANT) {
+              setSessionMessages((prev) => {
+                const newMessages = { ...prev };
+                const lastMsg =
+                  newMessages[currentConversationId]?.[
+                    newMessages[currentConversationId].length - 1
+                  ];
+                if (lastMsg && lastMsg.role === ROLE_ASSISTANT) {
                   lastMsg.error = t("chatInterface.requestTimeoutRetry");
                   lastMsg.isComplete = true;
                   lastMsg.thinking = undefined;
                 }
-                return lastMsg;
+                return newMessages;
               });
 
               if (currentConversationId && currentConversationId !== -1) {
@@ -438,9 +765,9 @@ export function ChatInterface() {
               log.error(t("chatInterface.errorCancelingRequest"), error);
             }
           }
-          uiState.conversationTimeoutsRef.current.delete(currentConversationId);
+          conversationTimeoutsRef.current.delete(currentConversationId);
         }, 120000);
-        uiState.conversationTimeoutsRef.current.set(currentConversationId, newTimeout);
+        conversationTimeoutsRef.current.set(currentConversationId, newTimeout);
       };
 
       // Before processing streaming response, set an initial timeout first
@@ -453,10 +780,10 @@ export function ChatInterface() {
         setCurrentSessionMessagesFactory(currentConversationId),
         resetTimeout,
         stepIdCounter,
-        uiState.setIsSwitchedConversation,
-        conversationManagement.isNewConversation,
-        conversationManagement.setConversationTitle,
-        conversationManagement.fetchConversationList,
+        setIsSwitchedConversation,
+        isNewConversation,
+        setConversationTitle,
+        fetchConversationList,
         currentConversationId,
         conversationService,
         false, // isDebug: false for normal chat mode
@@ -464,22 +791,22 @@ export function ChatInterface() {
       );
 
       // Reset all related states
-      uiState.setIsLoading(false);
-      uiState.setIsStreaming(false);
+      setIsLoading(false);
+      setIsStreaming(false);
 
       // Clean up controller and timeout for current conversation
-      uiState.conversationControllersRef.current.delete(currentConversationId);
-      const timeout = uiState.conversationTimeoutsRef.current.get(
+      conversationControllersRef.current.delete(currentConversationId);
+      const timeout = conversationTimeoutsRef.current.get(
         currentConversationId
       );
       if (timeout) {
         clearTimeout(timeout);
-        uiState.conversationTimeoutsRef.current.delete(currentConversationId);
+        conversationTimeoutsRef.current.delete(currentConversationId);
       }
 
       // Remove from streaming list (only when conversationId is not -1)
       if (currentConversationId !== -1) {
-        uiState.setStreamingConversations((prev) => {
+        setStreamingConversations((prev) => {
           const newSet = new Set(prev);
           newSet.delete(currentConversationId);
           return newSet;
@@ -487,9 +814,9 @@ export function ChatInterface() {
 
         // When conversation is completed, only add to completed conversation list when user is not in current conversation interface
         // Use ref to get the actual conversation the user is in
-        const currentUserConversation = uiState.currentSelectedConversationRef.current;
+        const currentUserConversation = currentSelectedConversationRef.current;
         if (currentUserConversation !== currentConversationId) {
-          uiState.setCompletedConversations((prev) => {
+          setCompletedConversations((prev) => {
             const newSet = new Set(prev);
             newSet.add(currentConversationId);
             return newSet;
@@ -502,45 +829,55 @@ export function ChatInterface() {
       // If user actively canceled, don't show error message
       const err = error as Error;
       if (err.name === "AbortError") {
-        uiState.updateLastMessageInConversation(currentConversationId, (lastMsg) => {
-          if (lastMsg.role === ROLE_ASSISTANT) {
+        setSessionMessages((prev) => {
+          const newMessages = { ...prev };
+          const lastMsg =
+            newMessages[currentConversationId]?.[
+              newMessages[currentConversationId].length - 1
+            ];
+          if (lastMsg && lastMsg.role === ROLE_ASSISTANT) {
             lastMsg.content = t("chatInterface.conversationStopped");
             lastMsg.isComplete = true;
             lastMsg.thinking = undefined; // Explicitly clear thinking state
           }
-          return lastMsg;
+          return newMessages;
         });
       } else {
         log.error(t("chatInterface.errorLabel"), error);
         // Show user-friendly error message instead of technical error details
         const errorMessage = t("chatInterface.errorProcessingRequest");
-        uiState.updateLastMessageInConversation(currentConversationId, (lastMsg) => {
-          if (lastMsg.role === ROLE_ASSISTANT) {
+        setSessionMessages((prev) => {
+          const newMessages = { ...prev };
+          const lastMsg =
+            newMessages[currentConversationId]?.[
+              newMessages[currentConversationId].length - 1
+            ];
+          if (lastMsg && lastMsg.role === ROLE_ASSISTANT) {
             lastMsg.content = errorMessage;
             lastMsg.isComplete = true;
             lastMsg.error = errorMessage;
             lastMsg.thinking = undefined; // Explicitly clear thinking state
           }
-          return lastMsg;
+          return newMessages;
         });
       }
 
-      uiState.setIsLoading(false);
-      uiState.setIsStreaming(false);
+      setIsLoading(false);
+      setIsStreaming(false);
 
       // Clean up controller and timeout for current conversation
-      uiState.conversationControllersRef.current.delete(currentConversationId);
-      const timeout = uiState.conversationTimeoutsRef.current.get(
+      conversationControllersRef.current.delete(currentConversationId);
+      const timeout = conversationTimeoutsRef.current.get(
         currentConversationId
       );
       if (timeout) {
         clearTimeout(timeout);
-        uiState.conversationTimeoutsRef.current.delete(currentConversationId);
+        conversationTimeoutsRef.current.delete(currentConversationId);
       }
 
       // Remove from streaming list (only when conversationId is not -1)
       if (currentConversationId !== -1) {
-        uiState.setStreamingConversations((prev) => {
+        setStreamingConversations((prev) => {
           const newSet = new Set(prev);
           newSet.delete(currentConversationId);
           return newSet;
@@ -548,9 +885,9 @@ export function ChatInterface() {
 
         // When conversation is completed, only add to completed conversation list when user is not in current conversation interface
         // Use ref to get the actual conversation the user is in
-        const currentUserConversation = uiState.currentSelectedConversationRef.current;
+        const currentUserConversation = currentSelectedConversationRef.current;
         if (currentUserConversation !== currentConversationId) {
-          uiState.setCompletedConversations((prev) => {
+          setCompletedConversations((prev) => {
             const newSet = new Set(prev);
             newSet.add(currentConversationId);
             return newSet;
@@ -560,8 +897,8 @@ export function ChatInterface() {
     } finally {
       // Only reset button states if we should (not when preprocessing fails)
       if (shouldResetButtonStates) {
-        uiState.setIsLoading(false);
-        uiState.setIsStreaming(false);
+        setIsLoading(false);
+        setIsStreaming(false);
       }
     }
   };
@@ -578,28 +915,33 @@ export function ChatInterface() {
     // Do not cancel any conversation requests, let them continue running in the background
 
     // Record current running conversation
-    if (uiState.streamingConversations.size > 0) {
+    if (streamingConversations.size > 0) {
       // Keep existing SSE connections active
     }
 
     // Reset all states
     setInput("");
-    uiState.setIsLoading(false);
-    uiState.setIsSwitchedConversation(false);
-    
-    // Use conversation management hook
-    conversationManagement.handleNewConversation();
-    uiState.setIsLoadingHistoricalConversation(false); // Ensure not loading historical conversation
+    setIsLoading(false);
+    setConversationId(-1);
+    setIsSwitchedConversation(false);
+    setConversationTitle(t("chatInterface.newConversation"));
+    setSelectedConversationId(null);
+
+    // Update ref to track current selected conversation
+    currentSelectedConversationRef.current = null;
+    setIsNewConversation(true); // Ensure set to new conversation state
+    setIsLoadingHistoricalConversation(false); // Ensure not loading historical conversation
 
     // Reset streaming state
-    uiState.setIsStreaming(false);
+    setIsStreaming(false);
 
     // Reset selected message and right panel state
-    uiState.setSelectedMessageId(undefined);
-    uiState.setShowRightPanel(false);
+    setSelectedMessageId(undefined);
+    setShowRightPanel(false);
 
     // Reset attachment state
-    attachmentHandlers.clearAttachments();
+    setAttachments([]);
+    setFileUrls({});
 
     // Clear URL parameters
     const url = new URL(window.location.href);
@@ -612,22 +954,42 @@ export function ChatInterface() {
     await new Promise((resolve) => setTimeout(resolve, 0));
 
     // Ensure new conversation scrolls to bottom
-    uiState.setShouldScrollToBottom(true);
+    setShouldScrollToBottom(true);
   };
 
+  const fetchConversationList = async () => {
+    try {
+      const dialogHistory = await conversationService.getList();
+      // Sort by creation time, newest first
+      dialogHistory.sort((a, b) => b.create_time - a.create_time);
+      setConversationList(dialogHistory);
+      return dialogHistory;
+    } catch (error) {
+      log.error(t("chatInterface.errorFetchingConversationList"), error);
+      throw error;
+    }
+  };
 
   // When switching conversation, automatically load messages
   const handleDialogClick = async (dialog: ConversationListItem) => {
     // When switching conversation, keep all SSE connections active
     // Do not cancel any conversation requests, let them continue running in the background
 
-    // Use conversation management hook
-    conversationManagement.handleConversationSelect(dialog);
-    uiState.setSelectedMessageId(undefined);
-    uiState.setShowRightPanel(false);
+    // Immediately set conversation state, avoid flashing new conversation interface
+    setSelectedConversationId(dialog.conversation_id);
+    setConversationId(dialog.conversation_id);
+    setConversationTitle(dialog.conversation_title);
+
+    // Update ref to track current selected conversation
+    currentSelectedConversationRef.current = dialog.conversation_id;
+    setSelectedMessageId(undefined);
+    setShowRightPanel(false);
+
+    // Set not new conversation state
+    setIsNewConversation(false);
 
     // When user views conversation, clear completed state
-    uiState.setCompletedConversations((prev) => {
+    setCompletedConversations((prev) => {
       const newSet = new Set(prev);
       newSet.delete(dialog.conversation_id);
       return newSet;
@@ -635,24 +997,24 @@ export function ChatInterface() {
 
     // Check if there are cached messages
     const hasCachedMessages =
-      messageManagement.hasMessages(dialog.conversation_id);
-    const isCurrentActive = dialog.conversation_id === conversationManagement.conversationId;
+      sessionMessages[dialog.conversation_id] !== undefined;
+    const isCurrentActive = dialog.conversation_id === conversationId;
 
     // Log: click conversation
     // If there are cached messages, ensure not to show loading state
     if (hasCachedMessages) {
-      const cachedMessages = messageManagement.getMessages(dialog.conversation_id);
+      const cachedMessages = sessionMessages[dialog.conversation_id];
       // If cache is empty array, force reload historical messages
       if (cachedMessages && cachedMessages.length === 0) {
-        uiState.setIsLoadingHistoricalConversation(true);
-        uiState.setIsLoading(true);
+        setIsLoadingHistoricalConversation(true);
+        setIsLoading(true);
 
         try {
           // Create new AbortController for current request
           const controller = new AbortController();
 
           // Set timeout timer - 120 seconds
-          uiState.timeoutRef.current = setTimeout(() => {
+          timeoutRef.current = setTimeout(() => {
             if (controller && !controller.signal.aborted) {
               try {
                 controller.abort(t("chatInterface.requestTimeout"));
@@ -660,11 +1022,11 @@ export function ChatInterface() {
                 log.error(t("chatInterface.errorCancelingRequest"), error);
               }
             }
-            uiState.timeoutRef.current = null;
+            timeoutRef.current = null;
           }, 120000);
 
           // Save current controller reference
-          uiState.abortControllerRef.current = controller;
+          abortControllerRef.current = controller;
 
           // Use controller.signal to make request with timeout
           const data = await conversationService.getDetail(
@@ -673,9 +1035,9 @@ export function ChatInterface() {
           );
 
           // Clear timeout timer after request completes
-          if (uiState.timeoutRef.current) {
-            clearTimeout(uiState.timeoutRef.current);
-            uiState.timeoutRef.current = null;
+          if (timeoutRef.current) {
+            clearTimeout(timeoutRef.current);
+            timeoutRef.current = null;
           }
 
           // Don't process result if request was canceled
@@ -713,24 +1075,31 @@ export function ChatInterface() {
             });
 
             // Update message array
-            messageManagement.setMessages(dialog.conversation_id, formattedMessages);
+            setSessionMessages((prev) => ({
+              ...prev,
+              [dialog.conversation_id]: formattedMessages,
+            }));
 
             // Clear any previous error for this conversation
-            conversationManagement.clearConversationLoadError(dialog.conversation_id);
+            setConversationLoadError((prev) => {
+              const newErrors = { ...prev };
+              delete newErrors[dialog.conversation_id];
+              return newErrors;
+            });
 
             // Asynchronously load all attachment URLs
-            loadAttachmentUrls(formattedMessages, dialog.conversation_id, messageManagement, t);
+            loadAttachmentUrls(formattedMessages, dialog.conversation_id);
 
             // Trigger scroll to bottom
-            uiState.setShouldScrollToBottom(true);
+            setShouldScrollToBottom(true);
 
             // Reset shouldScrollToBottom after a delay to ensure scrolling completes.
             setTimeout(() => {
-              uiState.setShouldScrollToBottom(false);
+              setShouldScrollToBottom(false);
             }, 1000);
 
             // Refresh history list
-            conversationManagement.fetchConversationList().catch((err) => {
+            fetchConversationList().catch((err) => {
               log.error(
                 t("chatInterface.refreshDialogListFailedButContinue"),
                 err
@@ -738,10 +1107,11 @@ export function ChatInterface() {
             });
           } else {
             // No longer empty cache, only prompt no history messages
-            conversationManagement.setConversationLoadErrorForId(
-              dialog.conversation_id,
-              t("chatStreamMain.noHistory") || "该会话无历史消息"
-            );
+            setConversationLoadError((prev) => ({
+              ...prev,
+              [dialog.conversation_id]:
+                t("chatStreamMain.noHistory") || "该会话无历史消息",
+            }));
           }
         } catch (error) {
           log.error(
@@ -751,21 +1121,24 @@ export function ChatInterface() {
           // if error, don't set empty array, keep existing state to avoid showing new conversation interface
           // Instead, we can show an error message or retry mechanism
 
-          conversationManagement.setConversationLoadErrorForId(dialog.conversation_id, "Failed to load conversation");
+          setConversationLoadError((prev) => ({
+            ...prev,
+            [dialog.conversation_id]: "Failed to load conversation",
+          }));
         } finally {
           // ensure loading state is cleared
-          uiState.setIsLoading(false);
-          uiState.setIsLoadingHistoricalConversation(false);
+          setIsLoading(false);
+          setIsLoadingHistoricalConversation(false);
         }
       } else {
         // Cache has content, display normally
-        uiState.setIsLoadingHistoricalConversation(false);
-        uiState.setIsLoading(false); // Ensure isLoading state is also reset
+        setIsLoadingHistoricalConversation(false);
+        setIsLoading(false); // Ensure isLoading state is also reset
 
         // For cases where there are cached messages, also trigger scrolling to the bottom.
-        uiState.setShouldScrollToBottom(true);
+        setShouldScrollToBottom(true);
         setTimeout(() => {
-          uiState.setShouldScrollToBottom(false);
+          setShouldScrollToBottom(false);
         }, 1000);
       }
     }
@@ -773,15 +1146,15 @@ export function ChatInterface() {
     // If there are no cached messages and not current active conversation, load historical messages
     if (!hasCachedMessages && !isCurrentActive) {
       // Set loading historical conversation state
-      uiState.setIsLoadingHistoricalConversation(true);
-      uiState.setIsLoading(true);
+      setIsLoadingHistoricalConversation(true);
+      setIsLoading(true);
 
       try {
         // Create new AbortController for current request
         const controller = new AbortController();
 
         // Set timeout timer - 120 seconds
-        uiState.timeoutRef.current = setTimeout(() => {
+        timeoutRef.current = setTimeout(() => {
           if (controller && !controller.signal.aborted) {
             try {
               controller.abort(t("chatInterface.requestTimeout"));
@@ -789,11 +1162,11 @@ export function ChatInterface() {
               log.error(t("chatInterface.errorCancelingRequest"), error);
             }
           }
-          uiState.timeoutRef.current = null;
+          timeoutRef.current = null;
         }, 120000);
 
         // Save current controller reference
-        uiState.abortControllerRef.current = controller;
+        abortControllerRef.current = controller;
 
         // Use controller.signal to make request with timeout
         const data = await conversationService.getDetail(
@@ -802,9 +1175,9 @@ export function ChatInterface() {
         );
 
         // Clear timeout timer after request completes
-        if (uiState.timeoutRef.current) {
-          clearTimeout(uiState.timeoutRef.current);
-          uiState.timeoutRef.current = null;
+        if (timeoutRef.current) {
+          clearTimeout(timeoutRef.current);
+          timeoutRef.current = null;
         }
 
         // Don't process result if request was canceled
@@ -842,24 +1215,31 @@ export function ChatInterface() {
           });
 
           // Update message array
-          messageManagement.setMessages(dialog.conversation_id, formattedMessages);
+          setSessionMessages((prev) => ({
+            ...prev,
+            [dialog.conversation_id]: formattedMessages,
+          }));
 
           // Clear any previous error for this conversation
-          conversationManagement.clearConversationLoadError(dialog.conversation_id);
+          setConversationLoadError((prev) => {
+            const newErrors = { ...prev };
+            delete newErrors[dialog.conversation_id];
+            return newErrors;
+          });
 
           // Asynchronously load all attachment URLs
-          loadAttachmentUrls(formattedMessages, dialog.conversation_id, messageManagement, t);
+          loadAttachmentUrls(formattedMessages, dialog.conversation_id);
 
           // Trigger scroll to bottom
-          uiState.setShouldScrollToBottom(true);
+          setShouldScrollToBottom(true);
 
           // Reset shouldScrollToBottom after a delay to ensure scrolling completes.
           setTimeout(() => {
-            uiState.setShouldScrollToBottom(false);
+            setShouldScrollToBottom(false);
           }, 1000);
 
           // Refresh history list
-          conversationManagement.fetchConversationList().catch((err) => {
+          fetchConversationList().catch((err) => {
             log.error(
               t("chatInterface.refreshDialogListFailedButContinue"),
               err
@@ -867,10 +1247,11 @@ export function ChatInterface() {
           });
         } else {
           // No longer empty cache, only prompt no history messages
-          conversationManagement.setConversationLoadErrorForId(
-            dialog.conversation_id,
-            t("chatStreamMain.noHistory") || "该会话无历史消息"
-          );
+          setConversationLoadError((prev) => ({
+            ...prev,
+            [dialog.conversation_id]:
+              t("chatStreamMain.noHistory") || "该会话无历史消息",
+          }));
         }
       } catch (error) {
         log.error(
@@ -880,24 +1261,72 @@ export function ChatInterface() {
         // if error, don't set empty array, keep existing state to avoid showing new conversation interface
         // Instead, we can show an error message or retry mechanism
 
-        conversationManagement.setConversationLoadErrorForId(dialog.conversation_id, "Failed to load conversation");
+        setConversationLoadError((prev) => ({
+          ...prev,
+          [dialog.conversation_id]: "Failed to load conversation",
+        }));
       } finally {
         // ensure loading state is cleared
-        uiState.setIsLoading(false);
-        uiState.setIsLoadingHistoricalConversation(false);
+        setIsLoading(false);
+        setIsLoadingHistoricalConversation(false);
       }
     }
   };
 
+  // Add function to asynchronously load attachment URLs
+  const loadAttachmentUrls = async (
+    messages: ChatMessageType[],
+    targetConversationId?: number
+  ) => {
+    // Create a copy to avoid directly modifying parameters
+    const updatedMessages = [...messages];
+    let hasUpdates = false;
+    const conversationIdToUse = targetConversationId || conversationId;
+
+    // Process attachments for each message
+    for (const message of updatedMessages) {
+      if (message.attachments && message.attachments.length > 0) {
+        // Get URL for each attachment
+        for (const attachment of message.attachments) {
+          if (attachment.object_name && !attachment.url) {
+            try {
+              // Get file URL
+              const url = await storageService.getFileUrl(
+                attachment.object_name
+              );
+              // Update attachment info
+              attachment.url = url;
+              hasUpdates = true;
+            } catch (error) {
+              log.error(
+                t("chatInterface.errorFetchingAttachmentUrl", {
+                  object_name: attachment.object_name,
+                }),
+                error
+              );
+            }
+          }
+        }
+      }
+    }
+
+    // If there are updates, set new message array
+    if (hasUpdates) {
+      setSessionMessages((prev) => ({
+        ...prev,
+        [conversationIdToUse]: updatedMessages,
+      }));
+    }
+  };
 
   // Left sidebar conversation title update
   const handleConversationRename = async (dialogId: number, title: string) => {
     try {
       await conversationService.rename(dialogId, title);
-      await conversationManagement.fetchConversationList();
+      await fetchConversationList();
 
-      if (conversationManagement.selectedConversationId === dialogId) {
-        conversationManagement.setConversationTitle(title);
+      if (selectedConversationId === dialogId) {
+        setConversationTitle(title);
       }
     } catch (error) {
       log.error(t("chatInterface.renameFailed"), error);
@@ -909,30 +1338,30 @@ export function ChatInterface() {
     try {
       // If deleting the currently active conversation, stop conversation first
       if (
-        conversationManagement.selectedConversationId === dialogId &&
-        uiState.isStreaming &&
-        conversationManagement.conversationId === dialogId
+        selectedConversationId === dialogId &&
+        isStreaming &&
+        conversationId === dialogId
       ) {
         // Cancel current ongoing request first
-        if (uiState.abortControllerRef.current) {
+        if (abortControllerRef.current) {
           try {
-            uiState.abortControllerRef.current.abort(
+            abortControllerRef.current.abort(
               t("chatInterface.deleteConversation")
             );
           } catch (error) {
             log.error(t("chatInterface.errorCancelingRequest"), error);
           }
-          uiState.abortControllerRef.current = null;
+          abortControllerRef.current = null;
         }
 
         // Clear timeout timer
-        if (uiState.timeoutRef.current) {
-          clearTimeout(uiState.timeoutRef.current);
-          uiState.timeoutRef.current = null;
+        if (timeoutRef.current) {
+          clearTimeout(timeoutRef.current);
+          timeoutRef.current = null;
         }
 
-        uiState.setIsStreaming(false);
-        uiState.setIsLoading(false);
+        setIsStreaming(false);
+        setIsLoading(false);
 
         try {
           await conversationService.stop(dialogId);
@@ -946,13 +1375,13 @@ export function ChatInterface() {
       }
 
       await conversationService.delete(dialogId);
-      await conversationManagement.fetchConversationList();
+      await fetchConversationList();
 
-      if (conversationManagement.selectedConversationId === dialogId) {
-        conversationManagement.setSelectedConversationId(null);
+      if (selectedConversationId === dialogId) {
+        setSelectedConversationId(null);
         // Update ref to track current selected conversation
-        conversationManagement.currentSelectedConversationRef.current = null;
-        conversationManagement.setConversationTitle(t("chatInterface.newConversation"));
+        currentSelectedConversationRef.current = null;
+        setConversationTitle(t("chatInterface.newConversation"));
         handleNewConversation();
       }
     } catch (error) {
@@ -960,45 +1389,70 @@ export function ChatInterface() {
     }
   };
 
-  // Handle image click preview - now handled by attachmentHandlers
+  // Add image error handling function
+  const handleImageError = (imageUrl: string) => {
+    log.error(t("chatInterface.imageLoadFailed"), imageUrl);
+
+    // Remove failed images from messages
+    setSessionMessages((prev) => {
+      const newMessages = { ...prev };
+      const lastMsg =
+        newMessages[conversationId]?.[newMessages[conversationId].length - 1];
+
+      if (lastMsg && lastMsg.role === ROLE_ASSISTANT && lastMsg.images) {
+        // Filter out failed images
+        lastMsg.images = lastMsg.images.filter((url) => url !== imageUrl);
+      }
+
+      return newMessages;
+    });
+  };
+
+  // Handle image click preview
+  const handleImageClick = (imageUrl: string) => {
+    setViewingImage(imageUrl);
+  };
 
   // Add conversation stop handling function
   const handleStop = async () => {
     // Stop agent_run of current conversation
     const currentController =
-      uiState.conversationControllersRef.current.get(conversationManagement.conversationId);
+      conversationControllersRef.current.get(conversationId);
     if (currentController) {
       try {
         currentController.abort(t("chatInterface.userManuallyStopped"));
       } catch (error) {
         log.error(t("chatInterface.errorCancelingRequest"), error);
       }
-      uiState.conversationControllersRef.current.delete(conversationManagement.conversationId);
+      conversationControllersRef.current.delete(conversationId);
     }
 
     // Clear timeout timer for current conversation
-    const currentTimeout = uiState.conversationTimeoutsRef.current.get(conversationManagement.conversationId);
+    const currentTimeout = conversationTimeoutsRef.current.get(conversationId);
     if (currentTimeout) {
       clearTimeout(currentTimeout);
-      uiState.conversationTimeoutsRef.current.delete(conversationManagement.conversationId);
+      conversationTimeoutsRef.current.delete(conversationId);
     }
 
     // Immediately update frontend state
-    uiState.setIsStreaming(false);
-    uiState.setIsLoading(false);
+    setIsStreaming(false);
+    setIsLoading(false);
 
     // If no valid conversation ID, just reset frontend state
-    if (!conversationManagement.conversationId || conversationManagement.conversationId === -1) {
+    if (!conversationId || conversationId === -1) {
       return;
     }
 
     try {
       // Call backend stop API - this will stop both agent run and preprocess tasks
-      await conversationService.stop(conversationManagement.conversationId);
+      await conversationService.stop(conversationId);
 
       // Manually update messages, clear thinking state
-      uiState.updateLastMessageInConversation(conversationManagement.conversationId, (lastMsg) => {
-        if (lastMsg.role === ROLE_ASSISTANT) {
+      setSessionMessages((prev) => {
+        const newMessages = { ...prev };
+        const lastMsg =
+          newMessages[conversationId]?.[newMessages[conversationId].length - 1];
+        if (lastMsg && lastMsg.role === ROLE_ASSISTANT) {
           lastMsg.isComplete = true;
           lastMsg.thinking = undefined; // Explicitly clear thinking state
 
@@ -1021,22 +1475,22 @@ export function ChatInterface() {
             }
           }
         }
-        return lastMsg;
+        return newMessages;
       });
 
       // remove from streaming list
-      uiState.setStreamingConversations((prev) => {
+      setStreamingConversations((prev) => {
         const newSet = new Set(prev);
-        newSet.delete(conversationManagement.conversationId);
+        newSet.delete(conversationId);
         return newSet;
       });
 
       // when conversation is stopped, only add to completed conversations list when user is not in current conversation interface
-      const currentUserConversation = uiState.currentSelectedConversationRef.current;
-      if (currentUserConversation !== conversationManagement.conversationId) {
-        uiState.setCompletedConversations((prev) => {
+      const currentUserConversation = currentSelectedConversationRef.current;
+      if (currentUserConversation !== conversationId) {
+        setCompletedConversations((prev) => {
           const newSet = new Set(prev);
-          newSet.add(conversationManagement.conversationId);
+          newSet.add(conversationId);
           return newSet;
         });
       }
@@ -1044,40 +1498,113 @@ export function ChatInterface() {
       log.error(t("chatInterface.stopConversationFailed"), error);
 
       // Optionally show error message
-      uiState.updateLastMessageInConversation(conversationManagement.conversationId, (lastMsg) => {
-        if (lastMsg.role === ROLE_ASSISTANT) {
+      setSessionMessages((prev) => {
+        const newMessages = { ...prev };
+        const lastMsg =
+          newMessages[conversationId]?.[newMessages[conversationId].length - 1];
+        if (lastMsg && lastMsg.role === ROLE_ASSISTANT) {
           lastMsg.isComplete = true;
           lastMsg.thinking = undefined; // Explicitly clear thinking state
           lastMsg.error = t(
             "chatInterface.stopConversationFailedButFrontendStopped"
           );
         }
-        return lastMsg;
+        return newMessages;
       });
     }
   };
 
   // Top title rename function
   const handleTitleRename = async (newTitle: string) => {
-    if (conversationManagement.selectedConversationId && newTitle !== conversationManagement.conversationTitle) {
+    if (selectedConversationId && newTitle !== conversationTitle) {
       try {
-        await conversationManagement.updateConversationTitle(conversationManagement.selectedConversationId, newTitle);
+        await conversationService.rename(selectedConversationId, newTitle);
+        await fetchConversationList();
+        setConversationTitle(newTitle);
       } catch (error) {
         log.error(t("chatInterface.renameFailed"), error);
       }
     }
   };
 
+  // Handle message selection
+  const handleMessageSelect = (messageId: string) => {
+    if (messageId !== selectedMessageId) {
+      // If clicking on new message, set as selected and open right panel
+      setSelectedMessageId(messageId);
+      // Auto open right panel
+      setShowRightPanel(true);
+    } else {
+      // If clicking on already selected message, toggle panel state
+      toggleRightPanel();
+    }
+  };
+
+  // Like/dislike handling
+  const handleOpinionChange = async (
+    messageId: number,
+    opinion: "Y" | "N" | null
+  ) => {
+    try {
+      await conversationService.updateOpinion({
+        message_id: messageId,
+        opinion,
+      });
+      setSessionMessages((prev) => {
+        const newMessages = { ...prev };
+        // Update the opinion_flag for the specific message in all conversations
+        Object.keys(newMessages).forEach((conversationId) => {
+          const messages = newMessages[parseInt(conversationId)];
+          if (messages) {
+            const messageIndex = messages.findIndex(
+              (msg) => msg.message_id === messageId
+            );
+            if (messageIndex !== -1) {
+              newMessages[parseInt(conversationId)] = [...messages];
+              newMessages[parseInt(conversationId)][messageIndex] = {
+                ...newMessages[parseInt(conversationId)][messageIndex],
+                opinion_flag: opinion || undefined,
+              };
+            }
+          }
+        });
+        return newMessages;
+      });
+    } catch (error) {
+      log.error(t("chatInterface.updateOpinionFailed"), error);
+    }
+  };
+
+  // Add event listener for conversation list updates
+  useEffect(() => {
+    const handleConversationListUpdate = () => {
+      fetchConversationList().catch((err) => {
+        log.error(t("chatInterface.failedToUpdateConversationList"), err);
+      });
+    };
+
+    window.addEventListener(
+      "conversationListUpdated",
+      handleConversationListUpdate
+    );
+
+    return () => {
+      window.removeEventListener(
+        "conversationListUpdated",
+        handleConversationListUpdate
+      );
+    };
+  }, []);
 
   return (
     <>
       <div className="flex h-screen">
         <ChatSidebar
-          conversationList={conversationManagement.conversationList}
-          selectedConversationId={conversationManagement.selectedConversationId}
-          openDropdownId={uiState.openDropdownId}
-          streamingConversations={uiState.streamingConversations}
-          completedConversations={uiState.completedConversations}
+          conversationList={conversationList}
+          selectedConversationId={selectedConversationId}
+          openDropdownId={openDropdownId}
+          streamingConversations={streamingConversations}
+          completedConversations={completedConversations}
           onNewConversation={handleNewConversation}
           onDialogClick={handleDialogClick}
           onRename={handleConversationRename}
@@ -1090,10 +1617,10 @@ export function ChatInterface() {
             router.push("/setup");
           }}
           onDropdownOpenChange={(open: boolean, id: string | null) =>
-            uiState.setOpenDropdownId(open ? id : null)
+            setOpenDropdownId(open ? id : null)
           }
-          onToggleSidebar={uiState.toggleSidebar}
-          expanded={uiState.sidebarOpen}
+          onToggleSidebar={toggleSidebar}
+          expanded={sidebarOpen}
           userEmail={user?.email}
           userAvatarUrl={user?.avatar_url}
           userRole={user?.role}
@@ -1103,47 +1630,47 @@ export function ChatInterface() {
           <div className="flex flex-1 overflow-hidden">
             <div className="flex-1 flex flex-col">
               <ChatHeader
-                title={conversationManagement.conversationTitle}
+                title={conversationTitle}
                 onRename={handleTitleRename}
               />
 
               <ChatStreamMain
                 messages={currentMessages}
                 input={input}
-                isLoading={uiState.isLoading}
+                isLoading={isLoading}
                 isStreaming={isCurrentConversationStreaming}
                 isLoadingHistoricalConversation={
-                  uiState.isLoadingHistoricalConversation
+                  isLoadingHistoricalConversation
                 }
                 conversationLoadError={
-                  conversationManagement.conversationLoadError[conversationManagement.selectedConversationId || 0]
+                  conversationLoadError[selectedConversationId || 0]
                 }
                 onInputChange={(value: string) => setInput(value)}
                 onSend={handleSend}
                 onStop={handleStop}
                 onKeyDown={handleKeyDown}
-                onSelectMessage={uiState.handleMessageSelect}
-                selectedMessageId={uiState.selectedMessageId}
-                onImageClick={attachmentHandlers.handleImageClick}
-                attachments={attachmentHandlers.attachments}
-                onAttachmentsChange={attachmentHandlers.handleAttachmentsChange}
-                onFileUpload={attachmentHandlers.handleFileUpload}
-                onImageUpload={attachmentHandlers.handleImageUpload}
-                onOpinionChange={uiState.handleOpinionChange}
-                currentConversationId={conversationManagement.conversationId}
-                shouldScrollToBottom={uiState.shouldScrollToBottom}
-                selectedAgentId={uiState.selectedAgentId}
-                onAgentSelect={uiState.setSelectedAgentId}
+                onSelectMessage={handleMessageSelect}
+                selectedMessageId={selectedMessageId}
+                onImageClick={handleImageClick}
+                attachments={attachments}
+                onAttachmentsChange={handleAttachmentsChange}
+                onFileUpload={handleFileUpload}
+                onImageUpload={handleImageUpload}
+                onOpinionChange={handleOpinionChange}
+                currentConversationId={conversationId}
+                shouldScrollToBottom={shouldScrollToBottom}
+                selectedAgentId={selectedAgentId}
+                onAgentSelect={setSelectedAgentId}
               />
             </div>
 
             <ChatRightPanel
               messages={currentMessages}
-              onImageError={uiState.handleImageErrorWrapper}
+              onImageError={handleImageError}
               maxInitialImages={14}
-              isVisible={uiState.showRightPanel}
-              toggleRightPanel={uiState.toggleRightPanel}
-              selectedMessageId={uiState.selectedMessageId}
+              isVisible={showRightPanel}
+              toggleRightPanel={toggleRightPanel}
+              selectedMessageId={selectedMessageId}
             />
           </div>
         </div>
@@ -1164,25 +1691,25 @@ export function ChatInterface() {
       </TooltipProvider>
 
       {/* Image preview */}
-      {attachmentHandlers.viewingImage && (
+      {viewingImage && (
         <div
           className="fixed inset-0 bg-black bg-opacity-70 flex items-center justify-center z-50"
-          onClick={attachmentHandlers.closeImageViewer}
+          onClick={() => setViewingImage(null)}
         >
           <div
             className="relative max-w-[90%] max-h-[90%]"
             onClick={(e) => e.stopPropagation()}
           >
             <img
-              src={attachmentHandlers.viewingImage}
+              src={viewingImage}
               alt={t("chatInterface.imagePreview")}
               className="max-w-full max-h-[90vh] object-contain"
               onError={() => {
-                uiState.handleImageErrorWrapper(attachmentHandlers.viewingImage!);
+                handleImageError(viewingImage);
               }}
             />
             <button
-              onClick={attachmentHandlers.closeImageViewer}
+              onClick={() => setViewingImage(null)}
               className="absolute -top-4 -right-4 bg-white p-1 rounded-full shadow-md hover:bg-white transition-colors"
               title={t("chatInterface.close")}
             >
