+++ frontend/next.config.mjs
@@ -21,9 +21,6 @@ const nextConfig = {
     parallelServerBuildTraces: true,
     parallelServerCompiles: true,
   },
-  // Optimize resource preloading
-  optimizeCss: true,
-  compress: true,
 }
 
 mergeConfig(nextConfig, userConfig)
